Remove unnecessary condition in fullScan
There are two reasons to remove this condition.
1. It's not used. All conditions when `attribute` is present are
already coverered via index lookup in other branches of switch in getProposal
(*a**,ea**,*av**,eav*,***n).
2. It's broken. Default branch of switch (which corresponds to FullScan) in
getProposal always populates `providing` in order of EAV, but if this
condition would ever run it would return values in order of AVE thus
breaking the assumptions of joinRound function.