Fix Chevrotain Lexer Initialization.
The second argument (boolean) to the constructor of a Chevrotain Lexer
should not be used in most productive flows.

It effectivily disables the validations performed on the Lexer.
For example:
* Finding duplicate patterns.
* Using Lexer Modes which have not been defined.
* and more...

This boolean argument defers the handling of the errors instead of just throwing an error.
This incorrect usage must have been copied / pasted from an example in the Chevrotain online playground
(Where the defering is used to display errors in the UI).