Update autocomplete sorting
Autocomplete now has the ability to sort matches by the alphabet as well as the default by length. The way this is done is by adding another value to your autocomplete called `sortby`. If `sortby` is equal to `alpha` then it sorts the matches by the alphabet. Not sure if this works 100%, modification probably needed.