Add string/split
I'm not sure if that is the right way to do this but maybe you could pass the results back in the form of a record. Not sure how to do that. Or maybe the user could choose the index they want to retrieve from the resulting array from splitting. Overall I think splitting strings would be helpful for when you've joined strings and need to unjoin them.